# OCI DIS Blueprint/apps/api/app/routers/exports.py
"""Exports router — synchronous artifact generation for M7."""

from fastapi import APIRouter, Depends
from fastapi.responses import FileResponse, Response
from sqlalchemy.ext.asyncio import AsyncSession

from app.core.db import get_db
from app.schemas.export import ExportJobResponse
from app.services import export_service

router = APIRouter(prefix="/exports", tags=["Exports"])


@router.get("/template/xlsx", summary="Download the offline capture template workbook")
async def download_capture_template(db: AsyncSession = Depends(get_db)) -> Response:
    workbook_bytes = await export_service.generate_capture_template(db)
    return Response(
        content=workbook_bytes,
        media_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
        headers={
            "Content-Disposition": (
                f"attachment; filename=oci-dis-import-template-v{export_service.TEMPLATE_VERSION}.xlsx"
            )
        },
    )


@router.post("/{project_id}/xlsx", response_model=ExportJobResponse, summary="Export catalog as XLSX")
async def export_xlsx(
    project_id: str,
    snapshot_id: str,
    db: AsyncSession = Depends(get_db),
) -> ExportJobResponse:
    return await export_service.create_xlsx_export(project_id, snapshot_id, db)


@router.get("/{project_id}/xlsx", summary="Download catalog export as XLSX using the latest snapshot")
async def download_xlsx(
    project_id: str,
    db: AsyncSession = Depends(get_db),
) -> FileResponse:
    snapshot_id = await export_service.latest_snapshot_id(project_id, db)
    job = await export_service.create_xlsx_export(project_id, snapshot_id, db)
    file_path, _ = export_service.get_export_file(project_id, job.job_id)
    return FileResponse(
        path=file_path,
        media_type="application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
        filename=job.filename,
    )


@router.get("/{project_id}/brief", summary="Download executive project brief as Markdown")
async def download_project_brief(
    project_id: str,
    db: AsyncSession = Depends(get_db),
) -> FileResponse:
    job = await export_service.create_brief_export(project_id, db)
    file_path, _ = export_service.get_export_file(project_id, job.job_id)
    return FileResponse(
        path=file_path,
        media_type="text/markdown; charset=utf-8",
        filename=job.filename,
    )


@router.post("/{project_id}/pdf", response_model=ExportJobResponse, summary="Export dashboard as PDF")
async def export_pdf(
    project_id: str,
    snapshot_id: str,
    db: AsyncSession = Depends(get_db),
) -> ExportJobResponse:
    return await export_service.create_pdf_export(project_id, snapshot_id, db)


@router.post("/{project_id}/json", response_model=ExportJobResponse, summary="Export project snapshot as JSON")
async def export_json(
    project_id: str,
    snapshot_id: str,
    db: AsyncSession = Depends(get_db),
) -> ExportJobResponse:
    return await export_service.create_json_export(project_id, snapshot_id, db)


@router.get("/{project_id}/jobs/{job_id}", response_model=ExportJobResponse, summary="Poll export job status")
async def get_export_job(project_id: str, job_id: str) -> ExportJobResponse:
    return await export_service.get_export_job(project_id, job_id)


@router.get("/{project_id}/jobs/{job_id}/download", summary="Download generated export artifact")
async def download_export(project_id: str, job_id: str) -> FileResponse:
    file_path, job = export_service.get_export_file(project_id, job_id)
    media_types = {
        "xlsx": "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
        "json": "application/json",
        "pdf": "application/pdf",
        "md": "text/markdown; charset=utf-8",
    }
    return FileResponse(
        path=file_path,
        media_type=media_types.get(job.format, "application/octet-stream"),
        filename=job.filename,
    )